requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
azure-cognitiveservices-speech==1.47.0
azure-core==1.36.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
fastapi==0.121.3
h11==0.16.0
httptools==0.7.1
idna==3.11
numpy==2.3.5
pandas==2.3.3
pillow==12.0.0
pydantic==2.12.4
pydantic-settings==2.12.0
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
six==1.17.0
sniffio==1.3.1
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
watchfiles==1.1.1
websockets==15.0.1
